requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
faiss-cpu
python-dotenv
streamlit
langchain_groq
langchain_core